Moves ExperimentCarousel component to experiments folder
